Move Header component in a dedicate folder
