+++ src/app/components/Header/Header.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './header.css';
 
 const Header = () => {
   return (
