+++ src/index.js
@@ -1,5 +1,4 @@
 import './app/components/css/body.css';
 import './app/components/css/app.css';
-import './app/components/css/header.css';
 
 import App from './app/components/App.jsx';
